Addresses #641, Failure when Word asks for citation on deleted item
I don't know this code at all, but here's a quick shot at fixing this problem. John, let me know if this is sufficient or if more needs to be done.

Simon, feel free to do something a bit more elegant (like clear missing items from the cache and the like)...