Fixes #769, Pages "is less than" in Advanced Search appears to never return results Addresses #775, Convert integers stored in DB as strings to integers
The Pages bug was actually a big broader and more complex. We were accidentally storing all values--including integer values--as strings, so the less-than operator didn't work on them (and the greater-than operator was matching most values, since "1"<"A").

Fixed by storing integers (up to 64-bits) not beginning with 0 as integers, casting numeric strings to integers for < and > comparisons, and using 'LIKE' instead of '=' for comparison of numeric search values for the "is"/"isNot" conditions. In 1.5, we'll handle #775.