Modifying agu.csl to conform to better conform to my understanding of AGU style based on their style guide and experience with AGU references.  Changes consist of:
1) Make it so that all references use initials for author's first
names.  I am not sure if this is officially in the style document, but
I have never seen an AGU publication that didn't use initials for
author names.  

2) Make it so that et al. is only used in bibliographic references
when you have more than 10 authors and that only the first author name
is used in this case, in accordance with
http://www.agu.org/pubs/AuthorRefSheet.pdf.

3) Set the disambiguate-add-givenname option to false for citations.
This MAY BE INCORRECT (not clear from above mentioned style guide - no
mention of what to do in ambiguous cases), but setting it to true
caused it to add the disambiguation in OpenOffice citations in wierd
places.  I think this might be a BUG in Zotero.  I am guessing that
upon citation Zotero was using disambiguate with respect to all
references in my database, not just those in the document, AND that
it considered names like John L. Doe and J. L. Doe different (or
something like that), causing it to add initials when it really wasn't
necessary.  Turning this off will produce the correct result in the
vaste majority of cases, but will fail in just those cases where
disambiguate is supposed to be used.