Making several minor changes to agu.csl to remove extra spaces that were being added to references.  I think these should work with most references, but I haven't checked with all possible reference types (for example books).
Also added a sort key for the bibliography based on the publication
year, but it is really hard to know if this is correct as sorting of
bibliographies in OpenOffice with the plugin seems to have bugs.

Also, added a sort by authors then year to the citation format.  I
believe this is correctly done and testing with the openoffice plugin
suggests that it is having some effect, though it isn't working quite
right (sometimes doesn't sort authors correctly - it seems to have
problems with et al. citations).