Fixes #901, downloadAssociatedFiles pref has no effect if automaticSnapshots is false
Cleaner version of zeppomedio's patch, with an additional check to not save links when automaticSnapshots is off (since that code would now be executed even if only downloadAssociatedFiles was checked)

I haven't tested this extensively, but the settings seem to now work as expected. Thanks to zeppomedio for the patch.