- Disabled some features (images, JS, META redirects, plugins) in the hidden browser -- this may affect some processDocuments()-based translators, but I'd like to see if it's possible to keep this stuff off. Might address #750, "Creating a hidden browser causes a persistent wait cursor", but I haven't checked. - Hidden browser wasn't being deleted when getCharsetFromFile() was used outside of attachment code - Fixed a couple JS strict errors
This addresses #1138, "dev xpi loads local files (and runs scripts in them) twice in the browser", but it doesn't fix it. The alert no longer appears due to the disabling of scripts, but getCharsetFromFile() is still being called when running detectWeb on local files. I assume this is due to the support for the URL bar importing of local files, but Simon, can we avoid calling getCharsetFromFile() in these cases, since we already have the charset in the (real, non-hidden) browser?