Fixes #1065, Dragging of attached files broken in Windows
Partially restores file dragging on Windows -- according to the Mozilla code, the method we were using to copy multiple files (and to properly copy snapshots as directories) is no longer supported on Windows in Firefox 3, but it's at least now possible to copy a single file out of Zotero. If the selected attachment is an HTML page, only the HTML file is copied, but that's probably better than nothing.

Also fixed drag and drop on Firefox 3.1 by adding support for the new HTML 5 Drag and Drop API.

In both old and new modes, dragging sometimes doesn't work properly and requires a restart--I'm not sure why.