- fix (most) unnecessary citation update issues - switch Zotero.JSON to native Firefox JSON support - update to citeproc-js 1.0.21
From Frank's announcement:
In this release:

- Tighten up internal "NUMERIC" update_mode to mean styles that render
citation-number in citations (renderings in the bibliography are now
ignored for purposes of setting this flag).

- The numeric styles fix introduced at version 1.0.17 broke with
styles that sort the bibliography on anything other than citation-
number (i.e. document first-reference order).  With this release,
arbitrary sorts of the bibliography work with numeric styles.

- Position evaluation code is now invoked only in styles that make use
of position testing (to save a few cycles).

- Numeric styles now perform targetted citation updates correctly.