update to citeproc-js 1.0.65
From Frank's release notes:

citeproc-js 1.0.64:

This release just pulls through the actual changes described in 1.0.62
and 1.0.63.

Apologies for any confusion this caused.

citeproc-js 1.0.65:

Do not needlessly circumvent name disambiguation logic. Reverses a
patch introduced in the 1.0.61 through 1.0.64 series of changes.

Apply the concise and correct fix for the undefined-in-name bug
previously addressed by the 1.0.61 through 1.0.64 series of releases.
Just replacing an undefined variable with an empty string does the
trick.

Force an initialized name that lacks an explicit string value on
initialize-with to full form, but only when using the by-cite
disambiguation rule.

Permit update of year-suffix when it is added, in addition to when it
exists and then changes.

Protect against an undefined value in updateItems(), which can
apparently sometimes occur in the course of word processor plugin
transactions.

Map page-first variable to page when snooping for a value. Prevents a
lone page-first cs:text node within a cs:group from being suppressed
for lack of output.

Repair recently amended appendCitationCluster() command to return a
list of paired values, in accordance with the processor manual.

Extend restoreProcessorState() to work with an empty argument, for
resetting the processor to a pristine state (for use case suggested by
Carles Pina).