Update to citeproc-js 1.0.68
From Frank's release notes:

Further refactoring of output queue mechanism, to unify all quote-
swapping and detection code for duplicate terminal punctuation and
spaces in a single function.  No changes to behavior, unless there is
a hole in the test suite somewhere.

Fix an output nesting mismatch error triggered by items with two
institutional authors in sequence, when rendered with a style that
sorts on the bibliography.

In the output queue mechanism, fix a bug in the migration of a leading
terminal character in a prefix to the preceding sibling suffix.