- update citeproc-js to 1.0.76 - fix issues with RTF export of brackets
From 1.0.71:
Variables (including names) should be (and have been) suppressed when
rendered via the cs:substitute node of cs:names.  Prior to this patch,
however, citeproc-js was unconditionally suppressing name variables
after their first use within a given cite, and this behavior differed
from the Zotero CSL 0.8.1 processor.  This patch restores the original
behavior, allowing names rendered outside of cs:substitute to be used
repeatedly within a cite.  (The only exception to this behavior is a
name rendered with et-al-min=1 and et-al-use-first=1, which will be
clipped from subsequent output within the same cite.)

The processor was using the localized close-inner-quote character for
apostrophes.  This caused problems in styles that use alternative
quotation marks.  The processor now uses a hard-coded apostrophe
character for apostrophes.

A bug in the normalization of names, for disambiguation purposes, that
contain punctuation or extraneous spaces has been fixed.

From 1.0.72:
Allow simple range, comma, and ampersand joins in the input to
cs:number, to permit "1st-2nd",  "xi-xv", etc.  When multiple numbers
are given, the following operations are performed: (1) ranges are
expanded and the numbers are converted to a list; (2) the resulting
list is sorted; (3) duplicates are eliminated; (4) ranges are
truncated; (5) numeric formatting is performed; and (6) appropriate
punctuation is applied to the result. (As of this writing, the use of
multiple numbers with cs:number is not covered by the CSL 1.0
specification, so this is an anticipatory change.  It is consistent
with the CSL 1.0 schema, however, and does not have any effect on
single input numbers. Production use of multiple number, issue,
volume, and edition variables should await the formal approval of a
CSL 1.0 schema change to allow the use of these variables with
cs:label, for contextual pluralization of any associated label.)

Eliminate space following a trailing apostrophe on a dropping-
particle, when immediately followed by the family name part in
rendered output.

Recognize multiple parameters to a position attribute on cs:if and
cs:else-if, per the CSL 1.0 schema.  For the fix, we eliminate cutted-
and-pasteded code from node_if.js and node_elseif.js, and consolidate
all position node-level position logic in attributes.js.

From 1.0.73:
Limit the recognition of multiple numbers to cases in which they are
separated by at least one space, to prevent the accidental
reformatting of things like "Document no. 1-3752,42".

From 1.0.74:
Recognize new Khmer locale for CSL provided by Hem Sras.

When parsing number ranges in the input to cs:number, abort parsing
when spread of range is over an arbitrary limit of 1000.

Support names-use-last attribute on cs:key. (Anticipatory change to
support upcoming change to CSL schema and specification.)

Use unicode ellipsis character rather than (nonexistent) locale term
when splicing author listings trucated with et-al-use-last.

From 1.0.75:
Do not apply decorative formatting to name suffixes.

Fix breakage in ordinal suffix formatting for days in a date.

For an all-lowercase name suffix (such as an explicit "et al."),
exclude the suffix from the name for disambiguation purposes, and
include it in the strict short form of the name in the short form of a
citation.

Add and recognize delimiter-precedes-et-al attribute, in line with
draft CSL proposal.

Implement grammatical gender sensitivity for ordinal suffixes, in line
with draft CSL proposal.

From 1.0.76:
Fix bug in locale processing, arising from the previous release, that
could trigger a processor crash.