update to citeproc-js 1.0.99
From 1.0.94:

The input format for multilingual data changes in this version.  The
change affects only multilingual data; ordinary input is not affected.

The previous multilingual input format, intended only for testing,
relied on string parsing to identify multilingual data embedded in a
given field. While this strategy worked for initial testing (its sole
intended purpose), it is cumbersome and potentially fragile.

The multilingual input data format recognized beginning with this
version of citeproc-js separates multilingual content properly into
separate elements, resulting in simpler design and more robust
operation.

The change to the multilingual input format is not backward
compatible; in the unlikely event that any  existing projects have
relied upon the old format, adjustments will be required to run this
version of the processor.

From 1.0.95:

Render multilingual names atomically, and do not modify the content of
names objects during processing. This permits implementation of
alternative name forms, needed to accomodate a new style rule in
scientific publishing, that permits author names to be given in the
original script alongside the romanized form.

Implement the rendering of original name forms, where an alternative
form of the name is used for the primary rendering.

From 1.0.96:

Fixes a global variable assignment. Important for stability.

From 1.0.97:

Replace the methods setLangTagsForCslNameTransliteration() and
setLangTagsForCslTitleTransliteration() with a single method
setLangTagsForCslTransliteration().

Change the multilingual parameters available through the default-
locale attribute of style files (parameters used primarily for
testing, not production) to reflect the API change described above.

Add method setOriginalCreatorNameFormatOption() to the control API.
This allows the name ordering and initialization conventions of
transliterated names either to track those of the original language,
or to follow the conventions of the target script.

Document new setOriginalCreatorNameFormatOption() method and  the
previously added method setOriginalCreatorNameFormsOption() in the
processor manual, and update the description of testing parameters
available via the default-locale attribute.

From 1.0.98:

Properly honor et-al-subsequent-min when it is smaller than et-al-min
(which it ordinarily is).

From 1.0.99:

Fix loop constraint variable misassignment that was a likely cause
behind the word processor plugin stability issue reported in the
following discussion thread:

http://forums.zotero.org/discussion/15777/error-updating-openofficeorg-document-in-21b/