Fixes #1804, Fix mysterious hang with citeproc-js 1.0.125
Re-merge citeproc-js 1.0.125 and use an alternative, slightly less efficient mechanism for reading from the integration pipe that shouldn't run into problems with garbage collection. Instead of trying to read from the pipe directly, we now spawn sh and cat to read from the pipe and send the output to a temporary file. We then read the temporary file. Because we don't block a JS thread, there are no GC issues. This works for Fx 4 as well, so I'm removing that code. I don't know whether there were potential GC issues in Fx 4, but if there were, I don't want to know about them.