Add a function to pump a generator until it yields false, allowing other events to be processed in between. This is useful because we can then call
yield true;

in place of Zotero.wait() to allow UI events to be processed without exiting the function, thus avoiding the hassle of setting up a large number of callbacks.

This is still painful compared to Zotero.wait(), since the yield has to be present in the generator passed to Zotero.pumpGenerator() and not a child function. However, it's less painful than using a bunch of nested setTimeout() calls.