Working version of relative paths (previous version modified the setter of attachmentPath which worked until the attachment path was saved to the SQL database and then pulled back out because save was using an absolute path). This version modifies Item.save() and the getter of Item.attachmentPath.  Item.s ave() now checks the attachment path to see if it is on the base attachment path  (set in preferences) and saves it in a marked form if it is.  The getter of att achmentPath checks the queried path and replaces the marked form with an absolut e path when appropriate.  The rest of Zotero deals with Item.attachmentPath and so should always see a full path just as it did without this patch.
I put the form for entering the base attachment path in the advanced tab of pref
erences.  I used the same method to set the path as used to set the data directo
ry.

I tried to place if statements in such a way that preferences, file paths, and the SQL database were not queried unless necessary (in case this matters for speed).  If the base attachment path preference is left set to , this patch should
 casue no change in the functionality of Zotero.