Ignore a file sync error on Windows with long ad junk files
For some reason on some Windows (XP, 7) systems a long path close to 255
characters long can pass destFile.create() and fail zipReader.extract().
This seems like a Mozilla bug, but until we find a better fix just
ignore the error for the only files where we've seen it, which are
advertising script artifacts with very long uninterrupted filenames
(which we should skip when saving anyway).

Also close zipReader after an extraction error