Change Zotero statute mapping from CSL bill to CSL legislation
A bill is a law that has not yet been enacted. Legislation is law
that has been enacted. The two are archived and cited differently,
and should be associated with different item types. CSL provides
a legislation type that is well suited to serve for laws that
have been enacted.

Among the existing styles in the repository, only two use
the legislation variable (american-association-of-petroleum-geologists
and irish-historical-studies). Both treat it as equivalent to bill. All
other styles use CSL bill exclusively.

It would be safe at this point to add legislation alongside bill
in all tests where it occurs (481 repository styles, apart from the
two mentioned above).

There is no particular urgency about this patch for official Zotero,
though, since statute support is likely to change considerably when
full legal support comes onstream anyway.