Upgrade citeproc-js to version 1.0.319
We recently introduced hard-wired code to apply hyphen as page
range delimiter in the French and Portuguese locales, with a view
to introducing a page-range-delimiter term in CSL 1.0.1. The new
term has been agreed within CSL, but official deployment must
await the CSL 1.0.1 schema release.

The original code was purely hard-wired. This change allows an
explicit page-range-delimiter term to be set within a style
(or in a locale file), as intended by the upcoming CSL schema.