Allow Verify Server to pass if uploaded file doesn't exist
With any storage service backed by S3 or another eventually consistent
storage service, there will be a short delay between when files
are uploaded and when they're available, so show a more helpful message
and allow the verification to pass. (Any IIS servers configured not to
serve extensionless files will also fail, but that's probably much
rarer at this point.)

We should also just retry the request on a 404, but that will have to
happen on master.

Addresses #115