Sort empty titles last when title isn't primary sort
When sorting by Title, empty titles get sorted to the top of the items
list for visibility, but when sorting by another column and using the
title as a secondary/tertiary sort, empty titles should get sorted last
so that new empty items go to the end of the list rather than the
middle.

This is a little weird, and the alternative would be to just always sort
empty titles last even when sorting by Title, but this preserves the
current behavior for Title sorting. (Before f0f6772b01a titles weren't
used for secondary sorting at all, so there's no precedent for
title-sorting behavior when sorting by another column.)

Addresses #275