Fix downloads on broken WebDAV servers
cloudme.com and maybe box.com appear to have stopped sending a
Last-Modified header for files, causing Zotero to skip the file download
step. Instead of relying on that, we now save a random id to the
lastsync file and just use that as an ETag.

Also migrate lastsync to lastsync.txt, which might be supported better
on some other broken WebDAV servers.