Restore panel auto-hide and increase show delay to 2 seconds
I think requiring everyone with Zotero installed to manually close the panel is
just too annoying. (Not allowing the auto-hide for a couple seconds after
opening to prevent accidental closes would be nice, but preventDefault() in the
panel's onpopuphiding() doesn't seem to work, nor does changing noautohide
after opening.)

Increase the delay to 2 seconds to try to show on top of the Firefox 28->29
upgrade panel on slower computers.