Add option to place DOI or URL as a link on the title of bibliographies
This depends on a hook available in citeproc-js from version 1.0.533.
The option appears in the "Create Bibliography from Items..." popup.
When links are set on the titles of a bibliography, printed output
of the DOI or URL (if any) is suppressed for readability.

The option takes effect in bibliographies created in HTML and in RTF
through the Zotero UI, but has no effect either in Quick Copy cites,
or in the word processor plugins.